Can we form another word from the first one?
